Added Zopfli-recompressed (TTF2WOFF) version
This saves space. If I turn *on* optimization, it proceeds to remove 228 bytes from the "name" table, which could potentially cause compatibility flag issues. It also cuts 98K from the glyf table, and I don't want that to break anything. Visual bugs are no fun. I did a table list to check if there's anything removed with the optimization turned off, and it doesn't appear to remove anything and the table byte sizes match the original WOFF from FontForge. Oh and even the FontForge tables are preserved (FFTM for instance). Also, it preserves the WOFF version flags I set, so in theory I can replace the current UnifontEX WOFF with this one, which is quite smaller than the original WOFF. It was already smaller than the WOFF2 version, but this is even smaller. It's not technically the limit (there are better forks of Zopfli) but it isn't currently achievable using existing code.